Simplify code (quadproduction/issues#200)
